+++ test/doing_util_test.rb
@@ -35,7 +35,7 @@ class DoingUtilTest < Test::Unit::TestCase
   end
 
   def test_format_time
-    item = @wwid.content[@wwid.current_section]['items'][0]
+    item = @wwid.content[@wwid.current_section][:items][0]
     interval = @wwid.get_interval(item, formatted: false, record: false)
     assert_equal(720, interval, 'Interval should match')
     minutes = interval / 60 % 60