+++ test/plugins/hooks.rb
@@ -6,10 +6,10 @@ Doing::Hooks.register :post_write do |filename|
   end
 end
 
-Doing::Hooks.register :post_read do |content|
+Doing::Hooks.register :post_read do |wwid|
   if ENV['HOOK_TEST']
     total = 0
-    content.each { |s, v| total += v['items'].count }
+    wwid.content.each { |s, v| total += v[:items].count }
     puts "Post read hook! Read #{total} items."
   end
 end